Update to fabric 1.21.3
The de.z0rdak.yawp.util.MessageUtil.sendMessage() update does not use
the "overlay" flag. I am unsure which way is desired.

The no-gliding rule is implemented as ending flight, not just preventing
taking off in a protected region, but both versions are included, with
the inactive one commented out.

The explosion handling is reworked to interpose a new ExplosionBehavior
that also checkes the region flags.