worldedit selection integration
Temporary until MarklerStick is fixed with visualization.

This is the initial integration with the WorldEdit selection mechanism.
/wp region <world> <region> info puts the YAWP region into the WE
selection. Subsequent WE commands such as //expand etc and WE tool uses
update both the WE selection and the YAWP region, live. The link is
broken on the first //desel command to WE.